requirements.txt
piexif
pillow
jsonpickle
atproto
smtplib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib